finally, separate into strings and ints
